Extract details into separate component
